requirements.txt
# Core ML
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Data handling
pandas>=2.0.0
h5py>=3.9.0

# Experiment tracking
wandb>=0.16.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Baselines
scikit-learn>=1.3.0
xgboost>=2.0.0

# Backtesting (Phase 6)
vectorbt>=0.26.0

# Testing & code quality
pytest>=7.4.0
ruff>=0.1.0
black>=23.7.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
